% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{brick_nc}
\alias{brick_nc}
\title{Converts a NetCDF file to a RasterBrick}
\usage{
brick_nc(file, layername = NULL)
}
\arguments{
\item{file}{a path to a NetCDF (ext = .nc) file}

\item{layername}{a string indicating the name of the layer if different from the filename}
}
\value{
a RasterBrick object
}
\description{
Takes NetCDF files from the Climatology Lab and converts them into
a RasterBrick object
}
