% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{define_seasons}
\alias{define_seasons}
\title{Associate brick layers with seasons}
\usage{
define_seasons(brick, layer_type = "month", hemisphere = "northern")
}
\arguments{
\item{brick}{a RasterBrick object}

\item{layer_type}{a character vector indicating if each layer represents a month or a year}

\item{hemisphere}{a character vector indicating if temperate northern or southen hemisphere seasons should be used}
}
\value{
a named list of numbers indicating which layer belongs to which time classification
}
\description{
Determine which raster layers correspond to which times of year based
on what each layer represents and which hemisphere to use; currently, only
temperate seasons are supported.
}
