% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{extract_data}
\alias{extract_data}
\title{Extract climate and weather variables}
\usage{
extract_data(
  brick,
  measures = c("mean"),
  timeframe = c("annual", "winter", "spring", "summer", "fall"),
  writefile = T,
  varname = NULL,
  layer_type = "month",
  hemisphere = "northern",
  directory = "./"
)
}
\arguments{
\item{brick}{an object of class RasterBrick}

\item{measures}{a character vector indicating how variables should be summarized (allows mean, min, max, and sum)}

\item{timeframe}{a character vector indicating the time over which variables should be summarized (allows annual, winter, spring, summer, fall, or wateryear)}

\item{writefile}{logical; if TRUE, saves a .csv file to the users' machine}

\item{varname}{a character vector indicating which climate or weather variable you are extracting (e.g. tmin, tmax)}

\item{layer_type}{a character vector indicating if each layer represents a month or a year}

\item{hemisphere}{a character vector indicating if temperate northern or southen hemisphere seasons should be used}

\item{directory}{a path to where files should be written if writefile=TRUE}
}
\value{
if writefile=TRUE, .csv files of matrices; if writefile=FALSE, a list of lists of matrices
}
\description{
Pulls climate or weather variables from a RasterBrick for
different time periods and summarizes them by simple metrics
}
