% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{generate_wget}
\alias{generate_wget}
\title{Generate a script to download NetCDF files}
\usage{
generate_wget(
  bbox,
  layers,
  startyear,
  endyear,
  db = "TerraClimate",
  writefile = TRUE,
  directory = "./"
)
}
\arguments{
\item{bbox}{an object of class extent indicating the bounding box, in latitude and longitude, to download (i.e. min and max lon and lat)}

\item{layers}{a character vector indicating which climate/weather layers to download}

\item{startyear}{an integer indicating the first year of data to download}

\item{endyear}{an integer indicating the last year of data to download}

\item{db}{a character vector indicating which database to use; currently only TerraClimate is supported}

\item{writefile}{logical; if TRUE, the resulting bash script is written locally}

\item{directory}{a path to a local directory where the bash script should be written if writefile=TRUE}
}
\value{
if writefile=TRUE, a .sh file in the specified directory; if writefile=FALSE, a character vector
}
\description{
Generate a bash script that can be run to download data from the
Climatology Lab. Currently only TerraClimate is supported.
}
\details{
If downloading TerraClimate data, layers should correspond to the 17 TerraClimate variables:
\describe{
\item{aet}{aet: Actual Evapotranspiration, monthly total (mm)}
\item{def}{def: Climate Water Deficit, monthly total (mm)}
\item{pet}{pet: Potential evapotranspiration, monthly total (mm)}
\item{ppt}{ppt: Precipitation, monthly total (mm)}
\item{q}{q: Runoff, monthly total (mm)}
\item{soil}{soil: Soil Moisture, total column - at end of month (mm)}
\item{srad}{srad: Downward surface shortwave radiation (W/m2)}
\item{swe}{swe: Snow water equivalent - at end of month (mm)}
\item{tmax}{tmax: Max Temperature, average for month (C)}
\item{tmin}{tmin: Min Temperature, average for month (C)}
\item{vap}{vap: Vapor pressure, average for month (kPa)}
\item{ws}{ws: Wind speed, average for month (m/s)}
\item{vpd}{vpd: Vapor Pressure Deficit, average for month (kpa)}
\item{PDSI}{PDSI: Palmer Drought Severity Index, at end of month (unitless)}
}
}
