% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{get_departures}
\alias{get_departures}
\title{Calculate climate departures}
\usage{
get_departures(
  vars,
  baseline.years,
  comparison.years,
  dist.method = c("mahalanobis", "sed"),
  yearbounds = NULL
)
}
\arguments{
\item{vars}{a list of matrices with equal dimensions containing climate or weather data; nrow is the number of cells or points and ncol is the number of years}

\item{baseline.years}{an integer vector of which years should be used for the historical period}

\item{comparison.years}{an integer vector of which years should be the future period relative to baseline}

\item{dist.method}{a character vector indicating if Mahalanobis ("mahalanobis") or
Standardized Euclidean ("sed") distance should be used}

\item{yearbounds}{a numeric vector of length 2 indicating the first and last year in the data}
}
\value{
a matrix of climate departures where each row is a cell or point and each column is a year in the comparison period
}
\description{
Calculates univariate or multivariate climate departures from
baseline conditions
}
\details{
This function makes all sorts of assumptions about your data, like that
each column represents an individual year and that those years are sequential
with the earliest year in the first column and no gaps between years.

If multiple variables are passed to the function (i.e. the list of vars has
length > 2), it calculates a multivariate departure; if only one variable is
passed to the function it returns a univariate departure.
}
