% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{get_trends}
\alias{get_trends}
\title{Calculate trends in variables}
\usage{
get_trends(dat, trend.type)
}
\arguments{
\item{dat}{a matrix where each row is a cell or site and each column is a year}

\item{trend.type}{a character vector indicating whether a Theil-Sen ("sen") or 'slope slope' ("ols") estimate should be returned}
}
\value{
a numeric vector equal to the number of rows (cells or sites) in dat
}
\description{
Calculates trends over time in climate and weather variables or climate departures
}
