% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{make_extent}
\alias{make_extent}
\title{Converts coordinates or polygons to bounding boxes}
\usage{
make_extent(lat = NULL, lon = NULL, polygon = NULL)
}
\arguments{
\item{lat}{a numeric or character vector of latitudes}

\item{lon}{a numeric or character vector of longitudes}

\item{polygon}{an object of class SpatialPolygons}
}
\value{
an object of class extent
}
\description{
If using a list of coordinates (e.g. study sites) or a polygon
(e.g. a species range map) to define the area of interest, this will convert
it to an extent
}
