% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{run_wget}
\alias{run_wget}
\title{Run a bash script to download NetCDF files}
\usage{
run_wget(script.path, directory = "./")
}
\arguments{
\item{script.path}{a bash file (ext = .sh) including the path to the file}

\item{directory}{a path to a directory indicating where results should be stored}
}
\value{
downloads NetCDF files to the users' machine
}
\description{
Given a bash script produced by generate_wget(), runs the script
and downloads results to the specified directory
}
