% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\docType{data}
\name{skeleton_wget}
\alias{skeleton_wget}
\title{A template for wget requests}
\format{
A character vector containing text as a scaffold for generate_wget()
}
\usage{
skeleton_wget
}
\description{
A template for wget requests
}
\keyword{datasets}
